python/efscape/requirements_dev.txt
atomicwrites==1.3.0
attrs==19.1.0
bumpversion==0.5.3
Click==7.0
click-log==0.3.2
devs==0.1.8
future==0.17.1
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
six==1.12.0
wcwidth==0.1.7
zeroc-ice==3.7.2
zipp==0.6.0
